requirements.txt
flask
playwright
gunicorn
pymongo
pytz
itsdangerous
Flask-Mail
dnspython
Flask-Dance
Authlib
Werkzeug
flask-talisman
pycryptodome
flask-cors
MLB-StatsAPI
pybaseball
python-slugify